// src/types/eden/OffLedger.ts
export const OffLedgerMany = t.Object({
  data: t.Array(OffLedgerOne),
  pagination: t.Object({
    total: t.Number(),
    limit: t.Number(),
    offset: t.Number(),
    next: t.Number(),
    prev: t.Number(),
  })
});